requirements.txt
langchain==0.1.6
langid==1.1.6
transformers==4.37.2
SpeechRecognition==3.10.1
torchaudio==2.2.0
torch==2.2.0
google-cloud-translate==3.15.1
soundfile==0.12.1
numpy==1.26.4
faiss-cpu==1.7.4
pygame==2.5.2
